[posix] utils: move to platform folder
