+++ cmake/treedata/android/subdirs.txt
@@ -2,14 +2,14 @@ xbmc/cores/RetroPlayer/process/android  cores/RetroPlayer/process/android
 xbmc/input/touch                        input/touch
 xbmc/input/touch/generic                input/touch/generic
 xbmc/filesystem/posix                   filesystem/posix
-xbmc/utils/posix                        utils/posix
 xbmc/windowing/android                  windowing/android
-xbmc/platform/posix                     posix
+xbmc/platform/posix                     platform/posix
+xbmc/platform/posix/utils               platform/posix/utils
 xbmc/platform/linux                     platform/linux
 xbmc/platform/linux/network             platform/linux/network
 xbmc/platform/linux/peripherals         platform/linux/peripherals
-xbmc/platform/android/activity          android/activity
-xbmc/platform/android/bionic_supplement android/bionicsupplement
+xbmc/platform/android/activity          platform/android/activity
+xbmc/platform/android/bionic_supplement platform/android/bionicsupplement
 xbmc/platform/android/network           platform/android/network
 xbmc/platform/android/peripherals       platform/android/peripherals
 xbmc/platform/android/powermanagement   platform/android/powermanagement
