+++ cmake/treedata/osx/subdirs.txt
@@ -1,4 +1,5 @@
 xbmc/platform/posix                   posix
+xbmc/platform/posix/utils             platform/posix/utils
 xbmc/platform/linux                   platform/linux
 xbmc/platform/linux/network           platform/linux/network
 xbmc/platform/darwin                  platform/darwin
@@ -8,7 +9,6 @@ xbmc/platform/darwin/osx/peripherals  platform/osx/peripherals
 xbmc/platform/darwin/osx/powermanagement platform/darwin/osx/powermanagement
 xbmc/platform/darwin/osx/storage      platform/osx/storage
 xbmc/filesystem/posix                 filesystem/posix
-xbmc/utils/posix                      utils/posix
 xbmc/windowing/osx                    windowing/osx
 xbmc/cores/RetroPlayer/process/osx    cores/RetroPlayer/process/osx
 xbmc/cores/VideoPlayer/Process/osx    cores/VideoPlayer/Process/osx
