+++ xbmc/platform/posix/utils/FileHandle.h
