+++ xbmc/platform/posix/utils/SharedMemory.h
