+++ xbmc/windowing/wayland/WindowDecorator.h
@@ -32,7 +32,7 @@
 #include "threads/CriticalSection.h"
 #include "Util.h"
 #include "utils/Geometry.h"
-#include "utils/posix/SharedMemory.h"
+#include "platform/posix/utils/SharedMemory.h"
 #include "WindowDecorationHandler.h"
 
 namespace KODI
